requirements.txt
pandas
psycopg2-binary
sshtunnel
python-dotenv
pytest
pytest-cov